+++ src/systempolicy/systemPolicy.go
@@ -32,6 +32,7 @@ import (
 )
 
 var log *zerolog.Logger
+var kubearmorRelayURL types.ConfigKubeArmorRelay
 
 func init() {
 	log = logger.GetInstance()
@@ -1467,7 +1468,13 @@ func DiscoverSystemPolicyMain() {
 func StartSystemLogRcvr() {
 	for {
 		if cfg.GetCfgSystemLogFrom() == "kubearmor" {
-			plugin.StartKubeArmorRelay(SystemStopChan, cfg.GetCfgKubeArmor())
+			err := plugin.StartKubeArmorRelay(SystemStopChan, cfg.GetCfgKubeArmor())
+			if val, ok := <-err; ok && val != nil {
+				url := cluster.GetKubearmorRelayURL()
+				kubearmorRelayURL.KubeArmorRelayURL = url
+				kubearmorRelayURL.KubeArmorRelayPort = "32767"
+				_ = plugin.StartKubeArmorRelay(SystemStopChan, kubearmorRelayURL)
+			}
 		} else if cfg.GetCfgSystemLogFrom() == "feed-consumer" {
 			fc.ConsumerMutex.Lock()
 			fc.StartConsumer()
